make subfolders for isis instruments
