+++ src/esssans/isis/sans2d/__init__.py
@@ -1,0 +1,15 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
+
+
+from . import data
+from . import masking
+
+providers = data.providers + masking.providers
+
+
+__all__ = [
+    'data',
+    'masking',
+    'providers',
+]
