+++ src/esssans/isis/zoom/__init__.py
@@ -1,0 +1,13 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
+
+
+from . import data
+
+providers = data.providers
+
+
+__all__ = [
+    'data',
+    'providers',
+]
